Moved new frameless window into Window header/source files
