+++ Applications/spire/include/spire/ui/title_bar.hpp
@@ -1,6 +1,5 @@
 #ifndef SPIRE_TITLE_BAR_HPP
 #define SPIRE_TITLE_BAR_HPP
-//#include <QAbstractNativeEventFilter>
 #include <QLabel>
 #include <QWidget>
 #include "spire/ui/ui.hpp"
@@ -55,12 +54,11 @@ namespace Spire {
       //! Returns the title text QLabel.
       // TODO: this is just a workaround for QTBUG-70873
       // https://bugreports.qt.io/browse/QTBUG-70873
+      // When updated, title bar will handle drag-moving the window using
+      // a nativeEventFilter
       QLabel* get_title_label() const;
 
       void update_window_flags();
-
-      //bool nativeEventFilter(const QByteArray& event_type, void* message,
-      //  long* result) override;
 
     protected:
       bool eventFilter(QObject* watched, QEvent* event) override;
