+++ Applications/spire/include/spire/ui/window.hpp
@@ -1,32 +1,31 @@
 #ifndef SPIRE_WINDOW_HPP
 #define SPIRE_WINDOW_HPP
-#include <QAbstractNativeEventFilter>
-#include <QWidget>
+#include <QMainWindow>
 #include "spire/ui/ui.hpp"
 
 namespace Spire {
 
   //! A customized window container for top-level windows.
-  class Window : public QWidget, QAbstractNativeEventFilter {
+  class Window : public QMainWindow {
     public:
 
       //! Constructs a Window.
-      /*!
-        \param body The widget displayed within the window.
-        \param parent The parent widget to the window.
+      /*
+        \param body The widget displayed within the window, and the window's
+                    parent.
       */
-      explicit Window(QWidget* body, QWidget* parent = nullptr);
+      explicit Window(QWidget* m_body);
 
       //! Sets the icon to display.
-      /*!
+      /*
         \param icon The icon to display when the window has focus.
       */
       void set_icon(const QImage& icon);
 
-      //! Sets the icon to display.
-      /*!
+      //! Sets the icons to display.
+      /*
         \param icon The icon to display when the window has focus.
-        \param icon The icon to display when the window lacks focus.
+        \param unfocused_icon The icon to display when the window lacks focus.
       */
       void set_icon(const QImage& icon, const QImage& unfocused_icon);
 
@@ -39,44 +38,29 @@ namespace Spire {
       //! Sets the icons to display using an SVG file with default sizes.
       /*
         \param icon_path Path to the SVG file to display when the window has
-               focus.
+                          focus.
         \param unfocused_icon_path Path to the SVG file to display when the
-               window lacks focus.
+                                   window lacks focus.
       */
       void set_svg_icon(const QString& icon_path,
         const QString& unfocused_icon_path);
 
-      bool nativeEventFilter(const QByteArray& event_type, void* message,
-        long* result) override;
+      void setFixedSize(int width, int height);
+
+      void setFixedSize(const QSize& size);
 
     protected:
-      bool event(QEvent* e) override;
-      bool eventFilter(QObject* watched, QEvent* event) override;
-      void mousePressEvent(QMouseEvent* event) override;
-      void mouseReleaseEvent(QMouseEvent* event) override;
-      void paintEvent(QPaintEvent* event) override;
+      void changeEvent(QEvent* event) override;
+      bool nativeEvent(const QByteArray &eventType, void *message,
+        long *result) override;
 
     private:
-      enum ActiveResizeRect {
-        NONE = 0,
-        TOP = 1,
-        RIGHT = 2,
-        BOTTOM = 4,
-        LEFT = 8,
-        TOP_LEFT = TOP | LEFT,
-        TOP_RIGHT = TOP | RIGHT,
-        BOTTOM_RIGHT = BOTTOM | RIGHT,
-        BOTTOM_LEFT = BOTTOM | LEFT
-      };
-      int m_current_active_rect;
-      QWidget* m_border;
-      QWidget* m_body;
+      QWidget* m_container;
       TitleBar* m_title_bar;
-      bool m_is_resizing;
+      int m_resize_area_width;
+      bool m_is_resizeable;
 
-      void handle_resize();
-      void set_border_stylesheet(const QColor& color);
-      void update_resize_cursor(const QPoint& pos);
+      void set_resizeable(bool resizeable);
   };
 }
