+++ Applications/spire/source/ui/window.cpp
@@ -1,74 +1,24 @@
 #include "spire/ui/window.hpp"
-#include <QApplication>
-#include <QDesktopWidget>
-#include <QIcon>
-#include <QLinearGradient>
-#include <QMouseEvent>
-#include <QPainter>
-#include <QRadialGradient>
-#include <QResizeEvent>
-#include <QVBoxLayout>
-#ifdef Q_OS_WIN
+#include <QEvent>
+#include <QGuiApplication>
+#include <QScreen>
 #include <dwmapi.h>
 #include <qt_windows.h>
-#endif
+#include <windowsx.h>
 #include "spire/spire/dimensions.hpp"
 #include "spire/ui/title_bar.hpp"
 
 using namespace Spire;
 
 namespace {
-  const auto PADDING_SIZE = 10;
-
-  const auto TOP_STOPS = QGradientStops({
-    QGradientStop(0.0, QColor(0, 0, 0, 43)),
-    QGradientStop(0.05, QColor(0, 0, 0, 15)),
-    QGradientStop(0.1, QColor(0, 0, 0, 8)),
-    QGradientStop(0.20, QColor(0, 0, 0, 4)),
-    QGradientStop(0.50, QColor(0, 0, 0, 1)),
-    QGradientStop(1, Qt::transparent)});
-
-  const auto BOTTOM_STOPS = QGradientStops({
-    QGradientStop(0.0, QColor(0, 0, 0, 43)),
-    QGradientStop(0.10, QColor(0, 0, 0, 15)),
-    QGradientStop(0.16, QColor(0, 0, 0, 8)),
-    QGradientStop(0.24, QColor(0, 0, 0, 4)),
-    QGradientStop(0.50, QColor(0, 0, 0, 1)),
-    QGradientStop(1, Qt::transparent)});
+  auto ICON_SIZE() {
+    return scale(26, 26);
+  }
 
   auto make_svg_window_icon(const QString& icon_path) {
     return imageFromSvg(icon_path, scale(26, 26),
       QRect(translate(8, 8), scale(10, 10)));
   }
-}
-
-Window::Window(QWidget* body, QWidget* parent)
-    : QWidget(parent),
-      m_body(body),
-      m_is_resizing(false) {
-  QWidget::window()->setAttribute(Qt::WA_TranslucentBackground);
-  setAttribute(Qt::WA_Hover);
-  resize(m_body->width(), m_body->height());
-  auto layout = new QVBoxLayout(this);
-  layout->setContentsMargins({scale_width(PADDING_SIZE),
-    scale_height(PADDING_SIZE), scale_width(PADDING_SIZE),
-    scale_height(PADDING_SIZE)});
-  layout->setSpacing(0);
-  m_border = new QWidget(this);
-  m_border->installEventFilter(this);
-  m_border->setObjectName("window_border");
-  m_border->resize(m_body->size() + scale(1, 1));
-  set_border_stylesheet("#A0A0A0");
-  layout->addWidget(m_border);
-  auto border_layout = new QVBoxLayout(m_border);
-  border_layout->setMargin(scale_width(1));
-  border_layout->setSpacing(0);
-  m_title_bar = new TitleBar(m_body, m_border);
-  border_layout->addWidget(m_title_bar);
-  border_layout->addWidget(m_body);
-  QWidget::window()->installEventFilter(this);
-  set_svg_icon(":icons/spire-icon-black.svg", ":icons/spire-icon-grey.svg");
-  qApp->installNativeEventFilter(this);
 }
 
 void Window::set_icon(const QImage& icon) {
@@ -89,260 +39,123 @@ void Window::set_svg_icon(const QString& icon_path,
     make_svg_window_icon(unfocused_icon_path));
 }
 
-#ifdef Q_OS_WIN
-bool Window::nativeEventFilter(const QByteArray& event_type, void* message,
-    long* result) {
-  auto msg = static_cast<::MSG*>(message);
-  auto handle = reinterpret_cast<::HWND>(QWidget::window()->effectiveWinId());
-  if(handle != msg->hwnd && !::IsChild(handle, msg->hwnd)) {
-    return false;
+void Window::setFixedSize(int width, int height) {
+  setFixedSize({width, height});
+}
+
+void Window::setFixedSize(const QSize& size) {
+  set_resizeable(false);
+  QMainWindow::setFixedSize(size);
+}
+
+void Window::set_resizeable(bool resizeable) {
+  m_is_resizeable = resizeable;
+  if(m_is_resizeable) {
+    setWindowFlags(windowFlags() | Qt::WindowMaximizeButtonHint);
+    auto hwnd = reinterpret_cast<HWND>(winId());
+    auto style = ::GetWindowLong(hwnd, GWL_STYLE);
+    ::SetWindowLong(hwnd, GWL_STYLE, style | WS_MAXIMIZEBOX | WS_THICKFRAME |
+      WS_CAPTION);
+  } else {
+    setWindowFlags(windowFlags() & ~Qt::WindowMaximizeButtonHint);
+    auto hwnd = reinterpret_cast<HWND>(winId());
+    auto style = ::GetWindowLong(hwnd, GWL_STYLE);
+    ::SetWindowLong(hwnd, GWL_STYLE, style & ~WS_MAXIMIZEBOX & ~WS_CAPTION);
   }
-  if(msg->message == WM_ACTIVATE) {
-    auto margins = ::MARGINS{-1, -1, -1, -1};
-    ::DwmExtendFrameIntoClientArea(msg->hwnd, &margins);
-    ::SetWindowPos(handle, nullptr, 0, 0, 0, 0, SWP_FRAMECHANGED |
-      SWP_NOZORDER | SWP_NOOWNERZORDER | SWP_NOMOVE | SWP_NOSIZE |
-      SWP_NOACTIVATE);
-    return true;
-  } else if((msg->message == WM_NCCALCSIZE) && msg->wParam == TRUE) {
-    *result = 0;
-    return true;
-  } else if(msg->message == WM_GETMINMAXINFO) {
-    auto maximize_dimensions = reinterpret_cast<::MINMAXINFO*>(msg->lParam);
-    auto rect = QApplication::desktop()->availableGeometry(this);
-    maximize_dimensions->ptMaxSize.x = rect.width();
-    maximize_dimensions->ptMaxSize.y = rect.height() - 1;
-    maximize_dimensions->ptMaxPosition.x = 0;
-    maximize_dimensions->ptMaxPosition.y = 0;
+  m_title_bar->update_window_flags();
+  const auto shadow = MARGINS{ 1, 1, 1, 1 };
+  DwmExtendFrameIntoClientArea(reinterpret_cast<HWND>(winId()),
+    &shadow);
+}
+
+void Window::changeEvent(QEvent* event) {
+  if(event->type() == QEvent::ActivationChange) {
+    if(isActiveWindow()) {
+      m_container->setStyleSheet("background-color: #A0A0A0;");
+    } else {
+      m_container->setStyleSheet("background-color: #C8C8C8;");
+    }
+  }
+}
+
+bool Window::nativeEvent(const QByteArray &eventType, void *message,
+    long *result) {
+  MSG* msg = reinterpret_cast<MSG*>(message);
+  if(msg->message == WM_NCCALCSIZE) {
     *result = 0;
     return true;
   } else if(msg->message == WM_NCHITTEST) {
-    *result = HTCLIENT;
-    return true;
-  } else if(msg->message == WM_SIZE) {
-    if(msg->wParam != SIZE_MAXIMIZED) {
-      auto g = QWidget::window()->frameGeometry();
-      auto region = ::CreateRectRgn(0, 0, g.width(), g.height());
-      ::SetWindowRgn(handle, region, FALSE);
-    } else {
-      ::SetWindowRgn(handle, NULL, FALSE);
-    }
-  } else if(msg->message == WM_NCPAINT) {
-    *result = 0;
-    return true;
-  } else if(msg->message == WM_NCACTIVATE) {
-    *result = TRUE;
-    return true;
-  } else if(msg->message == WM_SETTEXT) {
-    *result = TRUE;
-    return true;
-  }
-  return false;
-}
-#else
-bool Window::nativeEventFilter(const QByteArray& event_type, void* message,
-    long* result) {
-  return false;
-}
-#endif
-
-bool Window::event(QEvent* e) {
-  if(e->type() == QEvent::HoverMove) {
-    auto hover = static_cast<QHoverEvent*>(e);
-    if(m_is_resizing) {
-      handle_resize();
-    } else {
-      update_resize_cursor(hover->pos());
-    }
-    return true;
-  } else if(e->type() == QEvent::HoverLeave) {
-    m_current_active_rect = ActiveResizeRect::NONE;
-    setCursor(Qt::ArrowCursor);
-    return true;
-  } else if(e->type() == QEvent::WindowActivate) {
-    set_border_stylesheet("#A0A0A0");
-  } else if(e->type() == QEvent::WindowDeactivate) {
-    set_border_stylesheet("#C8C8C8");
-  }
-  return QWidget::event(e);
-}
-
-bool Window::eventFilter(QObject* watched, QEvent* event) {
-  if(watched == QWidget::window()) {
-    if(event->type() == QEvent::WindowStateChange) {
-      auto state_change_event = static_cast<QWindowStateChangeEvent*>(event);
-      if(!state_change_event->oldState().testFlag(Qt::WindowMaximized)
-          && QWidget::window()->isMaximized()) {
-        layout()->setContentsMargins({});
-      } else if(state_change_event->oldState().testFlag(Qt::WindowMaximized) &&
-          !QWidget::window()->isMaximized()) {
-        layout()->setContentsMargins(
-          {scale_width(PADDING_SIZE), scale_height(PADDING_SIZE),
-          scale_width(PADDING_SIZE), scale_height(PADDING_SIZE)});
+    auto window_rect = RECT{};
+    GetWindowRect(reinterpret_cast<HWND>(winId()), &window_rect);
+    auto x = GET_X_LPARAM(msg->lParam);
+    auto y = GET_Y_LPARAM(msg->lParam);
+    if(m_is_resizeable) {
+      if (x >= window_rect.left &&
+          x < window_rect.left + m_resize_area_width &&
+          y < window_rect.bottom &&
+          y >= window_rect.bottom - m_resize_area_width) {
+        *result = HTBOTTOMLEFT;
+        return true;
+      }
+      if (x < window_rect.right &&
+          x >= window_rect.right - m_resize_area_width &&
+          y < window_rect.bottom &&
+          y >= window_rect.bottom - m_resize_area_width) {
+        *result = HTBOTTOMRIGHT;
+        return true;
+      }
+      if (x >= window_rect.left &&
+          x < window_rect.left + m_resize_area_width &&
+          y >= window_rect.top &&
+          y < window_rect.top + m_resize_area_width) {
+        *result = HTTOPLEFT;
+        return true;
+      }
+      if(x < window_rect.right &&
+          x >= window_rect.right - m_resize_area_width &&
+          y >= window_rect.top &&
+          y < window_rect.top + m_resize_area_width) {
+        *result = HTTOPRIGHT;
+        return true;
+      }
+      if (x >= window_rect.left &&
+          x < window_rect.left + m_resize_area_width) {
+        *result = HTLEFT;
+        return true;
+      }
+      if (x < window_rect.right &&
+          x >= window_rect.right - m_resize_area_width) {
+        *result = HTRIGHT;
+        return true;
+      }
+      if (y < window_rect.bottom &&
+          y >= window_rect.bottom - m_resize_area_width) {
+        *result = HTBOTTOM;
+        return true;
+      }
+      if (y >= window_rect.top &&
+          y < window_rect.top + m_resize_area_width) {
+        *result = HTTOP;
+        return true;
       }
     }
-  }
-  return QWidget::eventFilter(watched, event);
-}
-
-void Window::mousePressEvent(QMouseEvent* event) {
-  if(event->button() == Qt::LeftButton) {
-    update_resize_cursor(event->pos());
-    if(m_current_active_rect != ActiveResizeRect::NONE) {
-      m_is_resizing = true;
+    QPoint pos = m_title_bar->mapFromGlobal({x, y});
+    if (m_title_bar->get_title_label()->geometry().contains(pos)) {
+      *result = HTCAPTION;
+      return true;
+    }
+    return false;
+  } else if(msg->message == WM_SIZE) {
+    if(msg->wParam == SIZE_MAXIMIZED) {
+      auto abs_pos = mapToGlobal(m_container->pos());
+      auto pos = QGuiApplication::screenAt(mapToGlobal({window()->width() / 2,
+        window()->height() / 2}))->geometry().topLeft();
+      pos = QPoint(std::abs(abs_pos.x() - pos.x()),
+        std::abs(abs_pos.y() - pos.y()));
+      setContentsMargins(pos.x(), pos.y(), pos.x(), pos.y());
+    } else if(msg->wParam == SIZE_RESTORED) {
+      setContentsMargins({});
     }
   }
-}
-
-void Window::mouseReleaseEvent(QMouseEvent* event) {
-  m_is_resizing = false;
-  update_resize_cursor(event->pos());
-}
-
-void Window::paintEvent(QPaintEvent* event) {
-  if(QWidget::window()->isMaximized()) {
-    return;
-  }
-  QPainter painter(this);
-  auto parent_size = geometry().size() -
-    QSize(2 * scale_width(PADDING_SIZE), 2 * scale_height(PADDING_SIZE));
-  auto shadow_size = scale(PADDING_SIZE, PADDING_SIZE);
-  auto right_start = shadow_size.width() + parent_size.width();
-  auto bottom_start = shadow_size.height() + parent_size.height();
-  QRect top_left_rect(QPoint(0, 0), shadow_size);
-  QRect top_rect(QPoint(shadow_size.width(), 0),
-    QSize(parent_size.width(), shadow_size.height()));
-  QRadialGradient top_left_gradient(top_left_rect.bottomRight(),
-    top_left_rect.width());
-  top_left_gradient.setStops(TOP_STOPS);
-  painter.fillRect(top_left_rect, top_left_gradient);
-  QLinearGradient top_gradient(top_rect.bottomLeft(), top_rect.topLeft());
-  top_gradient.setStops(TOP_STOPS);
-  painter.fillRect(top_rect, top_gradient);
-  auto top_right_rect = top_left_rect;
-  top_right_rect.translate(right_start, 0);
-  QRadialGradient top_right_gradient(top_right_rect.bottomLeft(),
-    top_right_rect.width());
-  top_right_gradient.setStops(TOP_STOPS);
-  painter.fillRect(top_right_rect, top_right_gradient);
-  auto bottom_left_rect = top_left_rect;
-  bottom_left_rect.translate(0, bottom_start);
-  QRadialGradient bottom_left_gradient(bottom_left_rect.topRight(),
-    bottom_left_rect.width());
-  bottom_left_gradient.setStops(BOTTOM_STOPS);
-  painter.fillRect(bottom_left_rect, bottom_left_gradient);
-  auto bottom_rect = top_rect;
-  bottom_rect.translate(0, bottom_start);
-  QLinearGradient bottom_gradient(bottom_rect.topLeft(),
-    bottom_rect.bottomLeft());
-  bottom_gradient.setStops(BOTTOM_STOPS);
-  painter.fillRect(bottom_rect, bottom_gradient);
-  auto bottom_right_rect = top_left_rect;
-  bottom_right_rect.translate(right_start, bottom_start);
-  QRadialGradient bottom_right_gradient(bottom_right_rect.topLeft(),
-    bottom_right_rect.width());
-  bottom_right_gradient.setStops(BOTTOM_STOPS);
-  painter.fillRect(bottom_right_rect, bottom_right_gradient);
-  QRect left_rect(QPoint(0, shadow_size.height()),
-    QSize(shadow_size.width(), parent_size.height()));
-  QLinearGradient left_gradient(left_rect.topRight(), left_rect.topLeft());
-  left_gradient.setStops(TOP_STOPS);
-  painter.fillRect(left_rect, left_gradient);
-  auto right_rect = left_rect;
-  right_rect.translate(right_start, 0);
-  QLinearGradient right_gradient(right_rect.topLeft(), right_rect.topRight());
-  right_gradient.setStops(BOTTOM_STOPS);
-  painter.fillRect(right_rect, right_gradient);
-  QWidget::paintEvent(event);
-}
-
-void Window::handle_resize() {
-  auto g = QWidget::window()->geometry();
-  auto p = QCursor::pos();
-  auto container_delta = g.size() - m_body->size();
-  auto max_size = m_body->maximumSize() + container_delta;
-  auto min_size = QSize(std::max(m_body->minimumWidth(), std::max(
-    m_body->layout()->minimumSize().width(),
-    std::max(m_title_bar->minimumWidth(),
-    m_title_bar->layout()->minimumSize().width()))), m_body->minimumHeight()) +
-    container_delta + QSize(4, 0);
-  if(m_current_active_rect == ActiveResizeRect::TOP ||
-      m_current_active_rect == ActiveResizeRect::TOP_LEFT ||
-      m_current_active_rect == ActiveResizeRect::TOP_RIGHT) {
-    auto min_y = g.bottom() - max_size.height() + 1;
-    auto max_y = g.bottom() - min_size.height() + 1;
-    g.setY(std::max(min_y, std::min(max_y, p.y())));
-  } else if(m_current_active_rect == ActiveResizeRect::BOTTOM ||
-      m_current_active_rect == ActiveResizeRect::BOTTOM_LEFT ||
-      m_current_active_rect == ActiveResizeRect::BOTTOM_RIGHT) {
-    auto min_y = g.top() + min_size.height() - 1;
-    auto max_y = g.top() + max_size.height() - 1;
-    g.setBottom(std::max(min_y, std::min(max_y, p.y())));
-  }
-  if(m_current_active_rect == ActiveResizeRect::LEFT ||
-      m_current_active_rect == ActiveResizeRect::TOP_LEFT ||
-      m_current_active_rect == ActiveResizeRect::BOTTOM_LEFT) {
-    auto min_x = g.right() - max_size.width() + 1;
-    auto max_x = g.right() - min_size.width() + 1;
-    g.setX(std::min(max_x, std::max(min_x, p.x())));
-  } else if(m_current_active_rect == ActiveResizeRect::RIGHT ||
-      m_current_active_rect == ActiveResizeRect::TOP_RIGHT ||
-      m_current_active_rect == ActiveResizeRect::BOTTOM_RIGHT) {
-    auto min_x = g.x() + min_size.width() - 1;
-    auto max_x = g.x() + max_size.width() - 1;
-    g.setRight(std::max(min_x, std::min(max_x, p.x())));
-  }
-  QWidget::window()->setGeometry(g);
-}
-
-void Window::set_border_stylesheet(const QColor& color) {
-  m_border->setStyleSheet(QString(R"(
-    #window_border {
-      border: %1px solid %3 %2px solid %3;
-    })").arg(scale_height(1)).arg(scale_width(1)).arg(color.name()));
-}
-
-void Window::update_resize_cursor(const QPoint& pos) {
-  m_current_active_rect = ActiveResizeRect::NONE;
-  if(!QWidget::window()->isMaximized() &&
-      (m_body->sizePolicy().horizontalPolicy() != QSizePolicy::Fixed ||
-      m_body->sizePolicy().verticalPolicy() != QSizePolicy::Fixed)) {
-    auto resize_width = scale_width(PADDING_SIZE);
-    if(pos.x() <= resize_width &&
-        m_body->sizePolicy().horizontalPolicy() != QSizePolicy::Fixed) {
-      m_current_active_rect |= ActiveResizeRect::LEFT;
-    } else if(pos.x() >= width() - resize_width &&
-        m_body->sizePolicy().horizontalPolicy() != QSizePolicy::Fixed) {
-      m_current_active_rect |= ActiveResizeRect::RIGHT;
-    }
-    auto resize_height = scale_height(PADDING_SIZE);
-    if(pos.y() <= resize_height &&
-        m_body->sizePolicy().verticalPolicy() != QSizePolicy::Fixed) {
-      m_current_active_rect |= ActiveResizeRect::TOP;
-    } else if(pos.y() >= height() - resize_height &&
-        m_body->sizePolicy().verticalPolicy() != QSizePolicy::Fixed) {
-      m_current_active_rect |= ActiveResizeRect::BOTTOM;
-    }
-  }
-  auto cursor_shape = [&] {
-    if(m_current_active_rect == ActiveResizeRect::TOP) {
-      return Qt::SizeVerCursor;
-    } else if(m_current_active_rect == ActiveResizeRect::TOP_RIGHT) {
-      return Qt::SizeBDiagCursor;
-    } else if(m_current_active_rect == ActiveResizeRect::RIGHT) {
-      return Qt::SizeHorCursor;
-    } else if(m_current_active_rect == ActiveResizeRect::BOTTOM_RIGHT) {
-      return Qt::SizeFDiagCursor;
-    } else if(m_current_active_rect == ActiveResizeRect::BOTTOM) {
-      return Qt::SizeVerCursor;
-    } else if(m_current_active_rect == ActiveResizeRect::BOTTOM_LEFT) {
-      return Qt::SizeBDiagCursor;
-    } else if(m_current_active_rect == ActiveResizeRect::LEFT) {
-      return Qt::SizeHorCursor;
-    } else if(m_current_active_rect == ActiveResizeRect::TOP_LEFT) {
-      return Qt::SizeFDiagCursor;
-    }
-    return Qt::ArrowCursor;
-  }();
-  setCursor(cursor_shape);
+  return QMainWindow::nativeEvent(eventType, message, result);
 }
